refactor : waitingFacade 삭제
