--- src/main/java/com/prgrms/catchtable/common/exception/ExceptionHandler.java
@@ -1,8 +1,0 @@
-package com.prgrms.catchtable.common.exception;
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ExceptionHandler {
-
-}
