+++ src/main/java/com/prgrms/catchtable/waiting/service/WaitingService.java
@@ -1,16 +1,20 @@
 package com.prgrms.catchtable.waiting.service;
 
 import static com.prgrms.catchtable.common.exception.ErrorCode.EXISTING_MEMBER_WAITING;
+import static com.prgrms.catchtable.common.exception.ErrorCode.NOT_EXIST_MEMBER;
+import static com.prgrms.catchtable.common.exception.ErrorCode.NOT_EXIST_SHOP;
 import static com.prgrms.catchtable.waiting.domain.WaitingStatus.PROGRESS;
 
 import com.prgrms.catchtable.common.exception.custom.BadRequestCustomException;
+import com.prgrms.catchtable.common.exception.custom.NotFoundCustomException;
 import com.prgrms.catchtable.member.domain.Member;
+import com.prgrms.catchtable.member.repository.MemberRepository;
 import com.prgrms.catchtable.shop.domain.Shop;
+import com.prgrms.catchtable.shop.repository.ShopRepository;
 import com.prgrms.catchtable.waiting.domain.Waiting;
 import com.prgrms.catchtable.waiting.dto.CreateWaitingRequest;
 import com.prgrms.catchtable.waiting.dto.CreateWaitingResponse;
 import com.prgrms.catchtable.waiting.dto.WaitingMapper;
-import com.prgrms.catchtable.waiting.facade.WaitingFacade;
 import com.prgrms.catchtable.waiting.repository.WaitingRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -24,11 +28,12 @@ public class WaitingService {
     private final LocalDateTime START_DATE_TIME = LocalDateTime.of(LocalDate.now(), LocalTime.of(0,0,0));
     private final LocalDateTime END_DATE_TIME = LocalDateTime.of(LocalDate.now(), LocalTime.of(23,59,59));
     private final WaitingRepository waitingRepository;
-    private final WaitingFacade waitingFacade;
+    private final MemberRepository memberRepository;
+    private final ShopRepository shopRepository;
     public CreateWaitingResponse createWaiting(Long shopId, CreateWaitingRequest request) {
         // 연관 엔티티 조회
-        Member member = waitingFacade.getMemberEntity(1L);
-        Shop shop = waitingFacade.getShopEntity(shopId);
+        Member member = getMemberEntity(1L);
+        Shop shop = getShopEntity(shopId);
 
         // shop 영업 중인지 검증
         shop.validateIfShopOpened(LocalTime.now());
@@ -55,5 +60,17 @@ public class WaitingService {
         if (waitingRepository.existsByMember(member)){
             throw new BadRequestCustomException(EXISTING_MEMBER_WAITING);
         }
+    }
+
+    public Member getMemberEntity(Long memberId){
+        return memberRepository.findById(memberId).orElseThrow(
+            ()-> new NotFoundCustomException(NOT_EXIST_MEMBER)
+        );
+    }
+
+    public Shop getShopEntity(Long shopId){
+        return shopRepository.findById(shopId).orElseThrow(
+            () -> new NotFoundCustomException(NOT_EXIST_SHOP)
+        );
     }
 }
